修改 攻击处理逻辑 新增 攻击力度判断

+++ src/main/kotlin/ac/github/oa/api/event/entity/OriginCustomDamageEvent.kt
@@ -10,7 +10,7 @@ class OriginCustomDamageEvent(
     val damager: Entity,
     val entity: Entity,
     var damage: Double,
-    val power: Double,
+    val power: Float,
     val attacker: LivingEntity?,
     val origin: EntityDamageByEntityEvent?
 ) : BukkitProxyEvent() {
